[AIC-py][eval] clean up file/folder structure
TODO (future PR): fix links in readme
                                                                                |
Test:

run the notebook and npx command
