+++ python/src/aiconfig/eval/api/__init__.py
@@ -1,0 +1,32 @@
+"""This file is officially the eval API.    
+
+Usage:
+
+from aiconfig.eval.api import (
+    brevity,
+    substring_match,
+    run_test_suite_with_inputs,
+    TestSuiteWithInputsSettings,
+)
+"""
+# pyright: reportWildcardImportFromLibrary=false
+from ..lib import (
+    TestSuiteWithInputsSettings,
+    run_test_suite_with_inputs,
+    run_test_suite_outputs_only,
+    SampleEvaluationFunction,
+)
+
+from ..metrics import (
+    brevity,
+    substring_match,
+)
+
+__all__ = [
+    "brevity",
+    "substring_match",
+    "run_test_suite_with_inputs",
+    "run_test_suite_outputs_only",
+    "TestSuiteWithInputsSettings",
+    "SampleEvaluationFunction",
+]
