+++ python/src/aiconfig/eval/examples/travel/travel_promptfoo_config.yaml
@@ -1,10 +1,10 @@
 prompts: ["{{question}}"]
 providers:
-  - exec:python -m 'aiconfig.eval.promptfoo.run_aiconfig' ./travel_aiconfig_test_suite_settings.json
+  - exec:python -m 'aiconfig.scripts.run_aiconfig' ./travel_aiconfig_test_suite_settings.json
 tests:
   - description: "Test if output contains the expected value, and measure the length of the output"
     vars:
-      question: Food
+      question: "different kinds of cuisines"
     assert:
       - type: python
         value: output.lower().find('magnolia bakery') != -1
@@ -12,7 +12,7 @@ tests:
         value: len(output) < 5000
   - description: "Test if output contains the expected value, and measure the length of the output"
     vars:
-      question: "Iconic midtown skyscrapers"
+      question: "iconic midtown skyscrapers"
     assert:
       - type: python
         value: output.lower().find('empire state building') != -1
