+++ python/src/aiconfig/eval/promptfoo/README.md
@@ -1,1 +1,1 @@
-Moved to python/src/aiconfig/eval/README.md.
+Moved to https://github.com/lastmile-ai/aiconfig/blob/main/python/src/aiconfig/eval/README.md
