--- python/src/aiconfig/eval/promptfoo/examples/travel/travel_parametrized.aiconfig.json
@@ -1,1 +1,0 @@
-../../../custom_eval/examples/travel/travel_parametrized.aiconfig.json
